Finalized documentation and pyth and dia adaptors

+++ tests/fuzzing/functional/FuzzDToken.sol
@@ -138,7 +138,7 @@ contract FuzzDToken is StatefulBaseMarket {
             )
         );
         require(borrowPossible);
-        (uint32 lastTimestampUpdated, , uint256 compoundRate) = DToken(dtoken)
+        (uint40 lastTimestampUpdated, , uint256 compoundRate) = DToken(dtoken)
             .marketData();
         require(lastTimestampUpdated + compoundRate > block.timestamp);
 
@@ -204,7 +204,7 @@ contract FuzzDToken is StatefulBaseMarket {
             )
         );
         require(borrowPossible);
-        (uint32 lastTimestampUpdated, , uint256 compoundRate) = DToken(dtoken)
+        (uint40 lastTimestampUpdated, , uint256 compoundRate) = DToken(dtoken)
             .marketData();
         require(lastTimestampUpdated + compoundRate <= block.timestamp);
